feat: Add Bootstrap and React Bootstrap dependencies
This commit adds Bootstrap and React Bootstrap as dependencies in the package.json file. This will allow us to use Bootstrap components in our React application.

The Bootstrap version has been updated to 5.3.3 and the React Bootstrap version has been updated to 2.10.4.

The NavBar component has been added to the project, which uses the React Bootstrap components to create a responsive navigation bar.