feat: Update NavBar component to use Bootstrap's navbar-toggler-icon
The NavBar component in the src/components/NavBar.js file has been updated to use Bootstrap's navbar-toggler-icon for the toggle button. This will improve the visual appearance and functionality of the navigation bar.

This change aligns with the recent addition of Bootstrap and React Bootstrap dependencies in the project, as mentioned in the previous commits.